# Conflicts:
#	lib/common/src/main/java/nl/tudelft/sem11b/data/models/UserModel.java
#	services/authentication/src/main/java/nl/tudelft/sem11b/authentication/GroupController.java
#	services/authentication/src/main/java/nl/tudelft/sem11b/authentication/services/GroupServiceImpl.java
#	services/authentication/src/main/java/nl/tudelft/sem11b/authentication/services/UserServiceImpl.java
#	services/authentication/src/test/java/nl/tudelft/sem11b/authentication/UserServiceTest.java
#	services/authentication/src/test/java/nl/tudelft/sem11b/authentication/services/GroupServiceTest.java

+++ services/admin/src/main/java/nl/tudelft/sem11b/admin/services/BuildingServiceImpl.java
@@ -26,7 +26,7 @@ public class BuildingServiceImpl implements BuildingService {
     }
 
     @Override
-    public Optional<BuildingModel> getBuilding(int id) {
+    public Optional<BuildingModel> getBuilding(long id) {
         return buildings.findById(id).map(Building::toModel);
     }
 }
